adding theme provider
