+++ src/App.tsx
@@ -3,7 +3,7 @@ import Home from "./pages/Home";
 import SignUp from "./pages/SignUp";
 import SignIn from "./pages/SignIn";
 import ForgotPassword from "./pages/ForgotPassword";
-import Theme from "theme";
+import Theme from "./theme";
 import styled from "styled-components";
 
 function App() {
@@ -15,12 +15,14 @@ function App() {
   </Theme>
   
   return (
-    <Routes>
-      <Route path="/" element={<Home />} />
-      <Route path="sign-up" element={<SignUp />} />
-      <Route path="sign-in" element={<SignIn />} />
-      <Route path="forgot-password" element={<ForgotPassword />} />
-    </Routes>
+      <ThemeProvider theme={theme}>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="sign-up" element={<SignUp />} />
+          <Route path="sign-in" element={<SignIn />} />
+          <Route path="forgot-password" element={<ForgotPassword />} />
+        </Routes>
+      </ThemeProvider>
   );
 }
