--- src/theme.js
@@ -1,29 +1,0 @@
-import React from "react";
-import { ThemeProvider } from "styled-components";
-
-export const theme = {
-    color: {
-        white: '#FFFFFF',
-        brightPurple: '#6610F2',
-        pink: '#D63384',
-        orange: '#FD7E14',
-        blue: '#0D6EFD',
-        darkGreen: '#198754',
-        slateGray: '#343A40',
-        darkBlue: '#242237',
-        lightGray: '#6C757D',
-        purple: '#6F42C1',
-        red: '#DC3545',
-        yellow: '#FFC107',
-        aquaBlue: '#0DCAF0',
-        seaGreen: '#20C997',
-        black: '#212529',
-        darkViolet: '#30304D'
-    }
-};
-
-const Theme = ({ children }) => (
-    <ThemeProvider theme={theme}>{children}</ThemeProvider>
-);
-
-export default Theme;
